+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 
+* Use tuple-merger backed select implementation on tarantool 2.10+ (it gives
+  less pressure on Lua GC).
+
 ## [0.9.0] - 20-10-21
 
 ### Added